rename folder and files of component
